requirements.txt
flask
pytesseract
tesseract
requests
pillow
fpdf
python-docx
docx